Fixes for compiler warnings: initialize variables, cast to avoid signed/ unsigned comparisons, undefine macros before possibly redefining them, etc.
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@581 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ eldorappi/PpiManager.cpp
@@ -39,7 +39,7 @@ bool PPIManager::newProduct(
     _productData[prodIndex] = p;
 
     // if we have a complete set, send them to the display
-    if (_currentProducts.size() == _nProducts) {
+    if (_currentProducts.size() == (unsigned int)_nProducts) {
         double startAng = elDegrees-0.45;
         double stopAng = elDegrees+0.45;
         if (startAng < 0.0) {
